move class to its own file
